Add test_utils to sierra_generator.
