+++ crates/sierra_generator/src/expr_generator_test.rs
@@ -1,47 +1,15 @@
-use defs::db::{AsDefsGroup, DefsDatabase, DefsGroup};
+use defs::db::DefsGroup;
 use defs::ids::{ExternFunctionLongId, FreeFunctionId, GenericFunctionId, LocalVarId, VarId};
-use filesystem::db::{AsFilesGroup, FilesDatabase, FilesGroup};
 use filesystem::ids::{CrateId, ModuleId};
-use parser::db::ParserDatabase;
 use pretty_assertions::assert_eq;
 use salsa::{InternId, InternKey};
-use semantic::db::{SemanticDatabase, SemanticGroup};
+use semantic::db::SemanticGroup;
 use semantic::ids::TypeId;
-use syntax::node::db::{AsSyntaxGroup, SyntaxDatabase, SyntaxGroup};
 
-use crate::db::{SierraGenDatabase, SierraGenGroup};
 use crate::expr_generator::generate_expression_code;
 use crate::expr_generator_context::ExprGeneratorContext;
 use crate::pre_sierra;
-
-#[salsa::database(
-    DefsDatabase,
-    SemanticDatabase,
-    SierraGenDatabase,
-    ParserDatabase,
-    SyntaxDatabase,
-    FilesDatabase
-)]
-#[derive(Default)]
-pub struct DatabaseImpl {
-    storage: salsa::Storage<DatabaseImpl>,
-}
-impl salsa::Database for DatabaseImpl {}
-impl AsFilesGroup for DatabaseImpl {
-    fn as_files_group(&self) -> &(dyn FilesGroup + 'static) {
-        self
-    }
-}
-impl AsSyntaxGroup for DatabaseImpl {
-    fn as_syntax_group(&self) -> &(dyn SyntaxGroup + 'static) {
-        self
-    }
-}
-impl AsDefsGroup for DatabaseImpl {
-    fn as_defs_group(&self) -> &(dyn defs::db::DefsGroup + 'static) {
-        self
-    }
-}
+use crate::test_utils::{replace_libfunc_ids, DatabaseImpl};
 
 fn generate_expr_code_for_test(
     db: &DatabaseImpl,
@@ -51,30 +19,6 @@ fn generate_expr_code_for_test(
     let mut expr_generator_context = ExprGeneratorContext::new(db, dummy_function_id);
     let (statements, res) = generate_expression_code(&mut expr_generator_context, block);
     (statements, res)
-}
-
-/// Replaces `ConcreteLibFuncId` with a dummy `ConcreteLibFuncId` whose debug string is the string
-/// representing the original `ConcreteLibFuncLongId`.
-/// For example, while the original debug string may be `[6]`, the resulting debug string may be
-/// `felt_const<2>`.
-fn replace_libfunc_ids(
-    db: &dyn SierraGenGroup,
-    statement: &pre_sierra::Statement,
-) -> pre_sierra::Statement {
-    match statement {
-        pre_sierra::Statement::SierraStatement(sierra::program::GenStatement::Invocation(p)) => {
-            pre_sierra::Statement::SierraStatement(sierra::program::GenStatement::Invocation(
-                sierra::program::GenInvocation {
-                    libfunc_id: db
-                        .lookup_intern_concrete_lib_func(p.libfunc_id.clone())
-                        .to_string()
-                        .into(),
-                    ..p.clone()
-                },
-            ))
-        }
-        _ => statement.clone(),
-    }
 }
 
 #[test]
