+++ crates/sierra_generator/src/lib.rs
@@ -4,4 +4,6 @@ pub mod expr_generator_context;
 pub mod id_allocator;
 pub mod pre_sierra;
 pub mod resolve_labels;
+#[cfg(test)]
+pub mod test_utils;
 pub mod utils;
